Adding Setup
- Majority of tests on Fey will be integrated. They must be executed sequentially.
- Setup should happen only once, right before all suites run
